+++ src/style.css
@@ -10,13 +10,13 @@ body {
   height: 70%;
   background-color: white;
   padding: 1em;
-  display: flex;
-  flex-direction: column;
   margin: 5em auto;
   width: 60%;
 }
 
-#list_elemnt {
+#list_elemnt,
+#input_div,
+.eachList {
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -24,22 +24,16 @@ body {
 
 #input_div {
   border: 2px solid rgb(231, 230, 230);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
 }
 
-#listObjId {
+#listObjId,
+.container {
   display: flex;
   flex-direction: column;
 }
 
 .eachList {
   border: 1px solid rgb(231, 230, 230);
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
 }
 
 #comp {
